Removed ROSH information
Users already provide detail about risk management on the 'purpose of placement' and 'Useful features of an AP' screen. We have not tested this with users, but believe the textbox might be confusing for users, as it's an empty textbox at the bottom of the OASys information. There's an assumption that this may be missed, or mistaken as an empty OASys field that's been pulled in. We're going to remove it for now and see how this tests in research.