Updated designs for 'Requesting further information'
We've designed screens for an assessor to request additional information. For MVP this will remain as an offline process. We have presented the PP details to the assessor on the screen, and added in a follow up page to enter the details.

Trello ticket - https://trello.com/c/utp8Bwd9