Updated move on screens
We've updated the move on screens to include guidance on the placement duration. We have a second version on Miro, but we'd like to test this first with users.

We've also included conditional reveals on the arrangement screens. If a user selects Yes or No they will need to provide details on

- what arrangements are in place (Yes)
- plans to make arrangements (No)