Updated access and healthcare screens
After chatting with Dan and Sarah, we think the question 'Does this person have care and support needs?' should be presented up front.

We've changed the additional info screen as we were concerned about the length of the text box. We've included a single text box for users to but all of the adjustments for the items selected.